requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
